菜单增删改查

+++ BasicsApi/Service/MenuService.cs
@@ -4,43 +4,34 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using BasicsApi.Dto;
 
 namespace BasicsApi.Service
 {
     public class MenuService
     {
-        private static readonly object padlock = new object();
-        private static WeixiaoSysContext db;
+        private  WeixiaoSysContext db;
         public MenuService(WeixiaoSysContext context)
         {
-            //if (db == null)
-            //{
-            //    lock (padlock)
-            //    {
-            //        if (db == null)
-            //        {
-            //            db = context;
-            //        }
-            //    }
-            //}
-            db = context;
+               db = new WeixiaoSysContext();
         }
         /// <summary>
-        /// 
+        ///
         /// </summary>
         /// <returns></returns>
         public async Task<List<Menu>> MenuList()
         {
-           return await db.Menu.ToListAsync();
+            return await db.Menu.ToListAsync();
         }
         /// <summary>
-        /// 
+        ///
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
         public async Task<Menu> MenuById(int id)
         {
-            return  await db.Menu.FirstOrDefaultAsync(o=>o.Id==id);
+            var data= await db.Menu.FirstOrDefaultAsync(o => o.Id == id);
+            return data;
         }
         /// <summary>
         /// 根据id获取它和子树
@@ -50,15 +41,32 @@ namespace BasicsApi.Service
         public async Task<List<Menu>> Menus(int? id)
         {
             var results = new List<Menu>();
-            results = await db.Menu.Where(o => o.Pid == id).ToListAsync();
+            results =await db.Menu.Where(o => o.Pid == id).ToListAsync();
             if (results.Count() == 0)
             {
                 return new List<Menu>();
             }
             foreach (var result in results)
             {
-                result.Children = await Menus(result.Id);
+                result.Children =await Menus(result.Id);
             }
+            return results;
+        }
+        public async Task<List<SelectDto>> SelectMenus(int? id)
+        {
+            var results = new List<SelectDto>();
+             var menus=await db.Menu.Where(o => o.Pid == id).Include(x => x.Children).ToListAsync();
+             foreach (var x in  menus)
+              {
+                  var dto = new SelectDto()
+                  {
+                      title = x.Text,
+                      label = x.Text,
+                      key = x.Id,
+                      children =await SelectMenus(x.Id)
+                  };
+                  results.Add(dto);
+              };
             return results;
         }
         public async Task<int> Add(Menu menu)
@@ -73,7 +81,10 @@ namespace BasicsApi.Service
         }
         public async Task<int> Delete(int id)
         {
-            var del = await db.Menu.FirstOrDefaultAsync(o => o.Id == id);
+            var del = await db.Menu.Include(m=>m.Children).FirstOrDefaultAsync(o => o.Id == id);
+            if(del.Children.Count>0){
+                throw new Exception("请先删除子菜单！");
+            }
             db.Menu.Remove(del);
             return await db.SaveChangesAsync();
         }
